requirements.txt
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
asyncpg==0.29.0
automata-lib==8.1.0
boto3==1.34.131
botocore
cached_method==0.1.0
click==8.1.7
fastapi==0.104.1
frozendict==2.3.10
greenlet==3.0.3
h11==0.14.0
idna==3.4
jmespath==1.0.1
joblib==1.3.2
networkx==3.2.1
numpy==1.26.4
pandas==2.2.2
psycopg2-binary==2.9.9
pydantic==2.5.1
pydantic_core==2.14.3
python-dateutil==2.9.0.post0
pytz==2024.1
s3transfer==0.10.1
scikit-learn==1.4.1.post1
scipy==1.12.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.30
starlette==0.27.0
threadpoolctl==3.4.0
typing_extensions==4.8.0
tzdata==2024.1
urllib3
uvicorn==0.24.0.post1
python-dotenv==0.19.1
xgboost==2.1.0
